+++ src/actions/actions.h
@@ -9,13 +9,13 @@ namespace todo {
 class Action {
 protected:
   Model *model_{nullptr};      ///< Model to perform actions on.
-  std::vector<u16> exe_path_;  ///< Path of execution.
+  std::vector<u64> exe_path_;  ///< Path of execution.
 
 public:
   /// \brief Parameterized constructor.
   /// \param model Reference to the model to work with.
   /// \param path Path (execution) of action.
-  Action(Model &model, std::vector<u16> path);
+  Action(Model &model, std::vector<u64> path);
 
   /// \brief Default destructor.
   virtual ~Action() = default;
@@ -29,14 +29,14 @@ public:
 
 class RemoveAction : public Action {
 private:
-  std::vector<u16> undo_path_{};  ///< Path of undo.
+  std::vector<u64> undo_path_{};  ///< Path of undo.
   Task task_{};                   ///< Removed task.
 
 public:
   /// \brief Parameterized constructor.
   /// \param model Reference to the model to work with.
   /// \param path Path (execution) of action.
-  RemoveAction(Model &model, std::vector<u16> path);
+  RemoveAction(Model &model, std::vector<u64> path);
 
   /// \brief execute action.
   virtual void execute() override;
@@ -47,7 +47,7 @@ public:
 
 class AddAction : public Action {
 private:
-  std::vector<u16> undo_path_{};  ///< Path of undo.
+  std::vector<u64> undo_path_{};  ///< Path of undo.
   Task task_{};                   ///< Added task.
 
 public:
@@ -55,7 +55,7 @@ public:
   /// \param model Reference to the model to work with.
   /// \param path Path (execution) of action.
   /// \param task Task to add.
-  AddAction(Model &model, std::vector<u16> path, Task task);
+  AddAction(Model &model, std::vector<u64> path, Task task);
 
   /// \brief execute action.
   virtual void execute() override;
@@ -74,7 +74,7 @@ public:
   /// \param model Reference to the model to work with.
   /// \param path Path (execution) of action.
   /// \param new_status The new status to be assigned to task.
-  StatusChangeAction(Model &model, std::vector<u16> path, Status new_status);
+  StatusChangeAction(Model &model, std::vector<u64> path, Status new_status);
 
   /// \brief execute action.
   virtual void execute() override;
@@ -93,7 +93,7 @@ public:
   /// \param model Reference to the model to work with.
   /// \param path Path (execution) of action.
   /// \param new_priority The new priority to be assigned to task.
-  PriorityChangeAction(Model &model, std::vector<u16> path, u16 new_priority);
+  PriorityChangeAction(Model &model, std::vector<u64> path, u16 new_priority);
 
   /// \brief execute action.
   virtual void execute() override;
